numbersRiddle.java

+++ numbersRiddle.java
@@ -14,8 +14,8 @@ public class numbersRiddle {
         double negativeDouble = -2.7;
 //-------------------------------------------------------------------------------------------------------------------------------------
         // Perform caclulations with test case 1
-        // initialize the tracking variable with the user's choice 
-        int currentTotal = positiveInteger;
+        // initialize the tracking variable with the test case 
+        double currentTotal = positiveInteger;
         
         // multiply the number by 2
         currentTotal *= 2; 
@@ -40,7 +40,7 @@ public class numbersRiddle {
 //-------------------------------------------------------------------------------------------------------------------------------------
 
         // Perform calculations with test case 2
-        // initialize the tracking variable with the user's choice 
+        // initialize the tracking variable with the test case
         int currentTotal = negativeInteger;
         
         // multiply the number by 2
@@ -66,7 +66,7 @@ public class numbersRiddle {
 //-------------------------------------------------------------------------------------------------------------------------------------
 
         // Perform calculations with test case 3
-        // initialize the tracking variable with the user's choice 
+        // initialize the tracking variable with the test case
         int currentTotal = zero;
         
         // multiply the number by 2
@@ -91,7 +91,7 @@ public class numbersRiddle {
 //-------------------------------------------------------------------------------------------------------------------------------------
 
         // Perform calculations with test case 4
-        // initialize the tracking variable with the user's choice 
+        // initialize the tracking variable with the test case
         int currentTotal = one;
         
         // multiply the number by 2
@@ -116,7 +116,7 @@ public class numbersRiddle {
 //-------------------------------------------------------------------------------------------------------------------------------------
 
         // Perform calculations with test case 5
-        // initialize the tracking variable with the user's choice 
+        // initialize the tracking variable with the test case 
         int currentTotal = positiveDouble;
         
         // multiply the number by 2
@@ -142,7 +142,7 @@ public class numbersRiddle {
 //-------------------------------------------------------------------------------------------------------------------------------------
 
         // Perform calculations with test case 6
-        // initialize the tracking variable with the user's choice 
+        // initialize the tracking variable with the test case 
         int currentTotal = negativeDouble;
         
         // multiply the number by 2
